templates/airflow/requirements.txt
azure-storage-blob
pytz
fake
requests
pandas
pyspark
kafka-python
ksql